Falling back to disabled toggle when client uuid is missing (#18)
Currently, the behaviour of a range condition is random picking a uuid when the client does not provide one.
For more stable behaviour, we propose to fallback to disabling the toggle in this case.
The consequence of this is that clients without uuids will not see the feature enabled unless it is forced on
or uses an always on condition. Currently, rolling out a feature for clients without uuid will require additional code changes
to be performed (i.e. changing the toggle default to always on). In the future, it should be possible to switch the toggle
unconditionally on from the backend server, and make rolling out features to all clients simpler again.